Ito na yung latest. -Jom Pakicheck nlang.

+++ app/Http/Controllers/OrderManagementController.php
@@ -368,11 +368,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function DeleteFlower_per_Bqt_SessionOrder($flower_ID)
 	{
-        /*if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			echo $flower_ID;
 			foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 				if($row->id == $flower_ID){
@@ -407,7 +407,7 @@ foreach($debtDetails as $debtDetails){
 
   public function DeleteAcessory_per_SessionBqt_Order($Acessory_ID)
 	{
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
         Session::put('loginSession','fail');
         return redirect() -> route('adminsignin');
     }
@@ -426,11 +426,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_ClearFlower_per_Bqt_Order($order_ID)
 	{
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			$AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
 			Cart::instance('OrderedBqt_Flowers')->destroy();
@@ -445,7 +445,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_Clear_Bqt_Order()
 	{
-		if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
         Session::put('loginSession','fail');
         return redirect() -> route('adminsignin');//
       }
@@ -462,7 +462,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_Clear_BqtSession_Order()
 	{
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
         Session::put('loginSession','fail');
         return redirect() -> route('adminsignin');
     }
@@ -483,7 +483,7 @@ foreach($debtDetails as $debtDetails){
 
     public function Clear_Cart()
     {
-     if(auth::check() == false){
+     if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -502,7 +502,7 @@ foreach($debtDetails as $debtDetails){
 
     public function Clear_Bouquet()
     {
-      if(auth::check() == false){
+      if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -520,7 +520,7 @@ foreach($debtDetails as $debtDetails){
 	public function saveCustomized_Bqt($order_ID)
 	{
 
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -585,7 +585,7 @@ foreach($debtDetails as $debtDetails){
 
 
 	public function saveNewCustomized_Bqt(){
-        if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
@@ -656,11 +656,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function ConfrimOrder()
 	{//
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 
 	          $cities = DB::table('cities')
 	          ->select('*')
@@ -678,7 +678,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function return_to_CreationOfOrder()
 	{//
-       if(auth::check() == false){
+       if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
